Fixed a reactions bug
fixed a bug where the bot was not allowed to remove reactions because of not having permissions. Now it removes only the reactions that he made.

Also added a verification to see if the bot is allowed to send messages in a channel. If not, it will not reply to the command, preventing crashes.